refactoring nearly completed

+++ src/import.sh
@@ -8,17 +8,29 @@ Usage
 
 Run it from the /src directory in the repo.
 
-./import.sh [test|prod]
+./import.sh [test|prod] [viewers] [content]
 ./import.sh
     Test mode
 
 ./import.sh prod
     Production mode
 
+Pass tasks to indicate what should be imported:
+
+viewers
+    Inspects the viewers.yaml file and the installed viewers in the
+    static directory in order to configure the supported viewers
+
+content
+    Imports content from the file system and fills mongodb tables
+    accordingly
 "
 
 flaskdebug=""
 flasktest=""
+
+doviewers="x"
+docontent="x"
 
 while [ ! -z "$1" ]; do
     if [[ "$1" == "--help" ]]; then
@@ -33,6 +45,10 @@ while [ ! -z "$1" ]; do
         flaskdebug="--debug"
         flasktest="test"
         shift
+    elif [[ "$1" == "viewers" ]]; then
+        doviewers="v"
+    elif [[ "$1" == "content" ]]; then
+        docontent="v"
     else
         echo "unrecognized argument '$1'"
         shift
@@ -43,18 +59,31 @@ done
 cd ..
 repodir="`pwd`"
 
-echo "Removing old data ..."
-rm -rf data
+if [[ "$doviewers" == "v" ]]; then
+    cd src
 
-echo "Copying fresh example data ..."
-cp -r exampledata data
+    export repodir
+    export flasktest
+    export flaskdebug
 
-cd src
+    echo "Entering viewer config in mongodb"
+    python3 pure3d/import.py viewers
+    echo "Done"
 
-export repodir
-export flasktest
-export flaskdebug
+elif [[ "$docontent" == "v" ]]; then
+    echo "Removing old data ..."
+    rm -rf data
 
-echo "Filling mongodb collections ..."
-python3 pure3d/import.py
-echo "Done"
+    echo "Copying fresh example data ..."
+    cp -r exampledata data
+
+    cd src
+
+    export repodir
+    export flasktest
+    export flaskdebug
+
+    echo "Filling mongodb collections ..."
+    python3 pure3d/import.py content
+    echo "Done"
+fi
